Add initial support for gathering analytics in the analysis server
I'm primarily interested in feedback on the general approach, which is
to have a single object to which all significant events are reported and
which is responsible for reporting or aggregating the data as
appropriate.

I removed a getter from AnalysisServer that should no longer be used,
though I haven't changed the code that directly references the
underlying analytics object so that existing code won't be impacted.

I also added a couple of references to the new class, but because we're
always creating a noop version there won't be any impact on users at
this point.

Once the overall approach has landed I'll start wiring up the existing
methods and adding new methods as appropriate.

Change-Id: I419f409c9df82fdc62a14128bb1f67723e9cf55f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/244983
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Phil Quitslund <pquitslund@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>