Add the hooks to capture analytics for requests and responses
In the process of testing this I found (and fixed) a couple of bugs
introduced in the recent conversion of the legacy code to handler
classes. I wanted to validate that I'd added hooks in all the right
places, so I wrote an AnalysisManager to keep track of request/response
pairs. It showed that some extra responses were being generated. I'm
still debating how best to add this to the normal tests to prevent
regressions, so you'll probably see it in another CL at some point.

Change-Id: Ib203c02ef27b6efb53a41a48696a3a39e61ed800
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/245260
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Danny Tuppeny <danny@tuppeny.com>