[cfe] Update inferred types through listeners
This CL changes the way inferred types are updated in the AST. Instead
of updating the AST directly, builders register a listener with the
OmittedTypeBuilder, that will be called when the inferred type has been
computed. This prepares for propagating inferred types to macro types
that are references to omitted types.

Change-Id: Ica5e1ed259b552d5de24000235f114ab1b9eb682
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/246642
Commit-Queue: Johnni Winther <johnniwinther@google.com>
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>