Fix Cider tests on Windows.
It is not very useful to run Cider tests on Windows.
And we could just disable these tests on Windows.

But this kind of tests might be useful to verify the state of the
system in DAS, when we un-fork Cider and DAS implementations.

Change-Id: I4a7ed641cd750b7626427cdab5104c533ea5407e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/247132
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Keerti Parthasarathy <keertip@google.com>