[analysis_server] Fix an edit conflict renaming folders containing files that reference siblings
Renaming a folder with files that referenced each other would fail with edit conflicts. The pass over the first file would update the reference to the other (going from "a.dart" to "../old/a.dart" as if the other file wasn't moving too) and then then when processing the second file, it would try to update the import back.

It doesn't seem to ever make sense to update relative paths when the "other" file is also in the folder being renamed, so this skips them.

Change-Id: I14c09977aa66017eecac5922d0e3caacf59350fb
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/247541
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>