[analysis_server] Always provide textEdit for LSP completions
Not providing textEdit means that the client will use insertText and compute its own range for what needs to be replaced. This may be incorrect if the name includes symbols that VS Code doesn't realise are part of the identifier.

To fix this, we should _always_ supply the range. This will result in a slightly larger payload, although an upcoming change to use itemDefaults will remove that for supporting clients.

Fixes https://github.com/Dart-Code/Dart-Code/issues/3943.

Change-Id: Ifd3f38ef96fecc6c745e00f511eb0bdbbd0d4a82
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/248143
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>