[cfe] Support omitted types in macros
This adds a DependentTypeBuilder used for type annotations in
generated macro code that refer to inferred types in the original
code.

These types are created using synthesized unique names in the
generated augmentation library, which are hardwired in the scope of
the augmentation library.

Change-Id: If28851614e8a844998d5ef27a132a616c0c4929d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/248123
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>