[cfe] Use InvalidExpression for targetless super access
This changes the encoding of super access without a target to use
InvalidExpression, similar to static access without a target, instead
of a SuperMethodInvocation/SuperPropertyGet/SuperPropertySet node with
a `null` interface target.

This is in preparation for using (Abstract)Super* nodes with the same
precision as Static*/Instance* nodes.

Change-Id: Ie0320c72241162a4c32a904a6608c9d29c2df401
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/248780
Reviewed-by: Jens Johansen <jensj@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>