[dart2js] Add deferred read/write to closed world and global analysis serialization.
Results after this CL:
Without the --deferred-serialization flag the linker phase finishes with ~10.2GB in ~120s.
With the --deferred-serialization flag the linker phase finishes with ~8.4GB in ~118s.
All other phases have negligible changes.

This is a ~17.6% reduction in memory usage and a small improvement in runtime.

Summary of this change:
With this change we target the new Deferrable concept to places that are triggering Kernel AST loads. By deferring the deserialization of these high-level data objects, we avoid the cascading deserialization of many other objects including much of the Kernel AST.

There are still a few locations that load the bodies of some functions. Some are slightly harder to defer for various reasons but represent a small number of function bodies.

This is a first pass and further improvements can probably be achieved with similar changes.

Change-Id: I7449f4222669a7ad8af0a6549d30117133aed03a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/247281
Reviewed-by: Joshua Litt <joshualitt@google.com>
Commit-Queue: Nate Biggs <natebiggs@google.com>