[dart2js] Change return type of `failedAt` to `Never`
This allows `e!` to be replaced by `e ?? failedAt(...)` where a more descriptive internal error is desired.

`failedAt` used to return `bool` so that you could write

    assert(condition || failedAt(...));

We have had assertions with messages for a long time, so that was changed to

    assert(condition, failedAt(...));

This still works with the type `Never`.

Change-Id: I132665d898a6aefe1fcf73235e88c156bab1e21e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/249730
Reviewed-by: Joshua Litt <joshualitt@google.com>
Reviewed-by: Nate Biggs <natebiggs@google.com>
Commit-Queue: Stephen Adams <sra@google.com>